backend-python/requirements.txt
fastapi
uvicorn
databases
asyncpg
pydantic
python-jose
PyJWT
